+++ crates/fuel-core/src/service/adapters/graphql_api/off_chain.rs
@@ -46,10 +46,7 @@ use fuel_core_types::{
     blockchain::{
         block::CompressedBlock,
         consensus::Consensus,
-        primitives::{
-            BlockAt,
-            BlockId,
-        },
+        primitives::BlockId,
     },
     entities::relayer::transaction::RelayedTransactionStatus,
     fuel_tx::{
@@ -153,10 +150,9 @@ impl OffChainDatabase for OffChainIterableKeyValueView {
 
     fn old_blocks(
         &self,
-        height: BlockAt,
+        height: Option<BlockHeight>,
         direction: IterDirection,
     ) -> BoxedIter<'_, StorageResult<CompressedBlock>> {
-        let height: Option<BlockHeight> = height.into();
         self.iter_all_by_start::<OldFuelBlocks>(height.as_ref(), Some(direction))
             .map(|r| r.map(|(_, block)| block))
             .into_boxed()