+++ crates/fuel-core/src/service/adapters/graphql_api/on_chain.rs
@@ -34,10 +34,7 @@ use fuel_core_types::{
     blockchain::{
         block::CompressedBlock,
         consensus::Consensus,
-        primitives::{
-            BlockAt,
-            DaBlockHeight,
-        },
+        primitives::DaBlockHeight,
     },
     entities::relayer::message::Message,
     fuel_tx::{
@@ -75,10 +72,9 @@ impl DatabaseBlocks for OnChainIterableKeyValueView {
 
     fn blocks(
         &self,
-        height: BlockAt,
+        height: Option<BlockHeight>,
         direction: IterDirection,
     ) -> BoxedIter<'_, StorageResult<CompressedBlock>> {
-        let height: Option<BlockHeight> = height.into();
         self.iter_all_by_start::<FuelBlocks>(height.as_ref(), Some(direction))
             .map(|result| result.map(|(_, block)| block))
             .into_boxed()